Chore:broom:: UseCase 분리
