+++ Projects/App/Sources/Domain/User/UserSignupUseCase.swift
@@ -1,8 +1,8 @@
 //
-//  UserUseCase.swift
+//  UsersignupUseCase.swift
 //  App
 //
-//  Created by 리아 on 2022/10/20.
+//  Created by Lee Myeonghwan on 2022/10/24.
 //  Copyright © 2022 zesty. All rights reserved.
 //
 
@@ -10,44 +10,13 @@ import Combine
 import Foundation
 import Network
 
-final class UserUseCase {
+final class UserSignupUseCase {
     
     // output
-    let isUserRegisteredSubject = PassthroughSubject<Bool, Never>()
     let isNickNameOverlapedSubject = PassthroughSubject<Bool, Never>()
     let isNickNameChangedSubject = PassthroughSubject<Bool, Never>()
     
     private var cancelBag = Set<AnyCancellable>()
-
-    func postSignUpUser() {
-        let userDTO = SignUpUserDTO(id: 1, email: "tmdgusya@suwon.ac.kr", organizationName: "수원대학교")
-
-        UserAPI.postSignUp(userDTO: userDTO)
-            .sink { error in
-                switch error {
-                case .failure(let error): print(error.localizedString)
-                case .finished: break
-                }
-            } receiveValue: { response in
-                print(response)
-            }
-            .store(in: &cancelBag)
-    }
-    
-    func postAccessTokenUser(accessToken: String) {
-        UserAPI.postAccessToken(accessToken: accessToken)
-            .sink { error in
-                switch error {
-                case .failure(let error): print(error.localizedString)
-                case .finished: break
-                }
-            } receiveValue: { [weak self] userOauthDTO in
-                guard let self = self else { return }
-                self.isUserRegisteredSubject.send(true)
-                UserDefaults.standard.authToken = userOauthDTO.authToken
-            }
-            .store(in: &cancelBag)
-    }
     
     func getNicknameValidationUser(nickname: String) {
         UserAPI.getNicknameValidation(nickname: nickname)
@@ -77,6 +46,21 @@ final class UserUseCase {
                 guard let self = self else { return }
                 UserDefaults.standard.userName = nickname
                 self.isNickNameChangedSubject.send(true)
+            }
+            .store(in: &cancelBag)
+    }
+    
+    func postSignUpUser() {
+        let userDTO = SignUpUserDTO(id: 1, email: "tmdgusya@suwon.ac.kr", organizationName: "수원대학교")
+
+        UserAPI.postSignUp(userDTO: userDTO)
+            .sink { error in
+                switch error {
+                case .failure(let error): print(error.localizedString)
+                case .finished: break
+                }
+            } receiveValue: { response in
+                print(response)
             }
             .store(in: &cancelBag)
     }
