+++ Projects/App/Sources/UI/User/Signup/ViewModel/NickNameInputViewModel.swift
@@ -11,7 +11,7 @@ import Foundation
 
 final class NickNameInputViewModel {
     
-    let useCase = UserUseCase()
+    let useCase = UserSignupUseCase()
     
     // Input
     @Published var nickNameText = ""
